+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 next version
 - start using a Changelog
 - add powershell as admin method and command
-
-
+- add safe start, stop and reboot def's
+- disply out from admin cmd's
+- default headless
 
 0.1.0
 - initial prototype with a log of functionality